#5 Rename submodule (vdif.reader → vdif.convert)
